Factor out browser-storage-base.tsx, add docs.
